sar-reportes v1.1.0: auto-update, paginación reg_ctas, badge fuente en reportes, fix uint8 numpy
- Auto-update: SarReportes detecta nueva versión en /api/version/agentes, descarga en background y muestra progreso en banner del menú
- Paginación client-side en reg_ctas: 500 filas/página, totales sobre todos los datos
- Badge LOCAL/REMOTO·agente en encabezado de todos los reportes
- Fix numpy uint8 out of bounds en filtro tipo D (campo LAPSO REG_CTAS)
- instalar_sar.ps1: -Cliente ahora obligatorio (antes default 'pilar')

+++ sar-reportes/agente.py
@@ -126,10 +126,11 @@ def _aplicar_filtros_numpy(arr, rec_size, campos_map, filtros):
             def _dval(s):
                 s = str(s).replace('-','')
                 return int(s) if len(s) == 8 else 0
-            yr = (arr[:,o  ]-48)*1000 + (arr[:,o+1]-48)*100 + (arr[:,o+2]-48)*10 + (arr[:,o+3]-48)
-            mn = (arr[:,o+4]-48)*10   + (arr[:,o+5]-48)
-            dy = (arr[:,o+6]-48)*10   + (arr[:,o+7]-48)
-            dint = yr.astype(np.int64)*10000 + mn.astype(np.int64)*100 + dy.astype(np.int64)
+            # cast a int64 ANTES de operar: uint8 no soporta multiplicadores > 255
+            yr = (arr[:,o  ].astype(np.int64)-48)*1000 + (arr[:,o+1].astype(np.int64)-48)*100 + (arr[:,o+2].astype(np.int64)-48)*10 + (arr[:,o+3].astype(np.int64)-48)
+            mn = (arr[:,o+4].astype(np.int64)-48)*10   + (arr[:,o+5].astype(np.int64)-48)
+            dy = (arr[:,o+6].astype(np.int64)-48)*10   + (arr[:,o+7].astype(np.int64)-48)
+            dint = yr*10000 + mn*100 + dy
             if isinstance(valor_f, dict):
                 if 'desde' in valor_f: mask &= dint >= _dval(valor_f['desde'])
                 if 'hasta' in valor_f: mask &= dint <= _dval(valor_f['hasta'])
